getNumberOfLeds returns int, fix warnings

+++ Software/src/Settings.cpp
@@ -40,7 +40,7 @@
 
 namespace
 {
-inline const WBAdjustment getLedAdjustment(size_t ledIndex)
+inline const WBAdjustment getLedAdjustment(int ledIndex)
 {
     using namespace SettingsScope;
 
@@ -1220,9 +1220,9 @@ void Settings::setMoodLampSpeed(int value)
 QList<WBAdjustment> Settings::getLedCoefs()
 {
     QList<WBAdjustment> result;
-    const size_t numOfLeds = getNumberOfLeds(getConnectedDevice());
+    const int numOfLeds = getNumberOfLeds(getConnectedDevice());
 
-    for(size_t led = 0; led < numOfLeds; ++led)
+    for (int led = 0; led < numOfLeds; ++led)
         result.append(getLedAdjustment(led));
 
     return result;
@@ -1610,9 +1610,9 @@ void Settings::migrateSettings()
 
             int remap[] = {3, 4, 2, 1, 0, 5, 6, 7, 8, 9};
 
-            size_t ledCount = getNumberOfLeds(SupportedDevices::DeviceTypeLightpack);
+            int ledCount = getNumberOfLeds(SupportedDevices::DeviceTypeLightpack);
             QMap<int, LedInfo> ledInfoMap;
-            for(size_t i = 0; i < ledCount; i++){
+            for (int i = 0; i < ledCount; i++){
                 LedInfo ledInfo;
                 ledInfo.isEnabled = isLedEnabled(i);
                 ledInfo.position = getLedPosition(i);
